requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.30.6
authlib==1.3.2
supabase==2.4.0
python-dotenv==1.0.1
Flask==3.0.3
flask-cors==4.0.1
authlib==1.3.2
supabase==2.4.0
python-dotenv==1.0.1

